Move project tests to correct folder
